update pareto result
