+++ bencher/example/example_pareto.py
@@ -46,6 +46,6 @@ This is a slightly unusual way of doing pareto optimisation as we are not using 
 
 if __name__ == "__main__":
     run_cfg_ex = bch.BenchRunCfg()
-    run_cfg_ex.repeats = 2
+    run_cfg_ex.repeats = 5
     run_cfg_ex.level = 2
     example_pareto(run_cfg_ex).report.show()
